extract registry container presenter into explorer common
